Bug fix hot reloading
